requirements.txt
streamlit>=1.32.0
pillow>=10.0.0
pandas>=2.0.0 